Refactor ignition schedules (#1407)
* Move start angle, end angle and channel
angle into IgnitionSchedule struct

* Remove endScheduleSetByDecoder

* Conditional compile of end tooth indicators

* Replace refreshIgnitionSchedule1 with
a call to adjustCrankAngle

* Use fixed width integers

* Remove IgnitionSchedule::endCompare

* Move dwell protection into scheduler.*

* Unit test overdwell protection

* Simplify overdwell protection.

* Doxygen comments

* Rename IgnitionSchedule members

* Simplify calculateIgnitionTrailingRotary

* Add checks to prevent negative angle deltas

* Set ignition schedules via angles
Moves the duration calculation calls inside the
scheduler code.

* Add angleToTimerTicks

* Remove unused ignition time out calcs

* Move sync matching to schedulers

* Move ignition scheduler sync state matching into scheduler.* (and unit test it)

* Hoist call to matchIgnitionSchedulersToSyncState() out of switch statement

* Simplify ignition angle calculation logic: no need to repeat the channel usage logic that init() already goes through.

* Move calculateIgnitionAngles() to scheduler.*

* Move setIgnitionChannels() into scheduler.*

* Improve unit test coverage of matchIgnitionSchedulersToSyncState

* Improve over dwell test coverage

* Move over dwell protection into ignition schedule controller

* Move setIgnitionChannel() into ignition controller

* Increase unit test coverage

* Make reset() a member function

* Move igniton schedule control functions into scheduler_ignition_controller.*

* Move ignition schedules to ignition schedule controller.

* Remove unused #defs

* Move setEndTeeth() call into calculateIgnitionAngles()

* Rename initialiseIgnitionSchedulers() -> resetIgnitionSchedulers()

* Move ignition schedule callback setup into ignition controller

* Remove duplicate code from matchIgnitionSchedulersToSyncState - replace with setCallbacks() call.

* initialiseIgnitionSchedules() calls resetIgnitionSchedulers()

* setCallbacks() is part of the init path and therefore is not performance critical

+++ speeduino/globals.cpp
@@ -38,8 +38,8 @@ volatile uint8_t compositeLogHistory[TOOTH_LOG_SIZE];
 volatile unsigned int toothHistoryIndex = 0; ///< Current index to @ref toothHistory array
 unsigned long currentLoopTime; /**< The time (in uS) that the current mainloop started */
 volatile uint16_t ignitionCount; /**< The count of ignition events that have taken place since the engine started */
-int CRANK_ANGLE_MAX_IGN = 360;
-int CRANK_ANGLE_MAX_INJ = 360; ///< The number of crank degrees that the system track over. Typically 720 divided by the number of squirts per cycle (Eg 360 for wasted 2 squirt and 720 for sequential single squirt)
+int16_t CRANK_ANGLE_MAX_IGN = 360;
+int16_t CRANK_ANGLE_MAX_INJ = 360; ///< The number of crank degrees that the system track over. Typically 720 divided by the number of squirts per cycle (Eg 360 for wasted 2 squirt and 720 for sequential single squirt)
 volatile uint32_t runSecsX10;
 volatile uint32_t seclx10;
 volatile byte HWTest_INJ = 0; /**< Each bit in this variable represents one of the injector channels and it's HW test status */
